wraethdao/jb-712-project-description-has-html-tags-in-it-in-project-cards
